feat(subscription): support to fallback to pull subscription using 'proxy'
